router: always set MessageContext in getRequest
Otherwise, early errors like "empty payload" or json.Unmarshal errors
would not be sent to the user, since SendError would silently do nothing
when request.MessageContext was nil.

In that unwanted edge case, SendError now loudly errors:

	ERROR	router/router.go:188	cannot reply with error as MessageContext==nil: empty payload

This was a pretty severe error, as the HTTP server does not close a
client's connection until a response is sent via the router. And, since
the router failed to send the error response, the client in this
scenario remained waiting until the timeout was hit.

This no longer is the case with the blind-ca server:

	$ curl -X POST http://127.0.0.1:5000/ca
	{"response":{"error":"empty payload","ok":false,...